Add "editor.multiCursorPaste": "full" setting.
Defaults to "spread", which works as I'd expect until you have the same number of cursors as the number of lines copied (copied as one block, though). In that case, say, 2 cursors and one block of 2 lines copied and "spread" as the setting, pasting will put the first line from the copied block at cursor 1 and the second line at cursor 2.